+++ app/src/main/java/com/devYoussef/nb3elkhieradmin/model/OrderResponse.kt
@@ -18,7 +18,7 @@ data class OrderResponse(
         val products: List<Product?>?,
         val promoCode: PromoCode?,
         val quantity: Int?,
-        val totalPrice: Int?,
+        val totalPrice: Double?,
         val updatedAt: String?,
         val userId: UserId?
     ) {
@@ -39,8 +39,8 @@ data class OrderResponse(
                 val isOffered: Boolean?,
                 val name: String?,
                 val offer: List<Offer?>?,
-                val originalPrice: Int?,
-                val price: Int?,
+                val originalPrice: Double?,
+                val price: Double?,
                 val priceCurrency: String?,
                 val quantity: Int?,
                 val shortDescription: String?,
@@ -54,7 +54,7 @@ data class OrderResponse(
                 data class Offer(
                     val _id: String?,
                     val itemNum: Int?,
-                    val priceOffered: Int?
+                    val priceOffered: Double?
                 )
             }
         }
@@ -65,7 +65,7 @@ data class OrderResponse(
             val code: String?,
             val createdAt: String?,
             val isActive: Boolean?,
-            val price: Int?,
+            val price: Double?,
             val timesNum: Int?,
             val updatedAt: String?,
             val usedNum: Int?,
